requirements.txt
#django stuff
django==4.0.1
gunicorn==20.1.0
djangorestframework==3.13.1
django-filter==21.1

# other packages
biopython==1.79
django-environ=0.8.1
psycopg2-binary==2.9.3

